+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - removed 'position' from event properties
 - renamed vertex.getPosition in vertex.getIndex
 - renamed vertex.remove in vertex.delete
+- added 'editable:drawing:cancel' event
+- removed 'editable:edited' in favor of 'editable:drawing:finish'
 
 ## 0.1.0
 - initial release